refactor(gui): react-doctor cleanup for Dashboard
