+++ gui/src/pages/dashboard-overview-panels.tsx
@@ -1,0 +1,22 @@
+import MemoryObservabilityCard from "../components/MemoryObservabilityCard";
+import type { useDashboardData } from "./use-dashboard-data";
+import {
+  DashboardEffortCapPanel,
+  DashboardInjectionPanel,
+  DashboardMaintenancePanel,
+  DashboardSidecarPanels,
+} from "./dashboard-overview-sections";
+
+type Dash = ReturnType<typeof useDashboardData>;
+
+export function DashboardOverviewPanels(props: Dash) {
+  return (
+    <>
+      <DashboardEffortCapPanel apiBase={props.apiBase} d={props} />
+      <DashboardInjectionPanel apiBase={props.apiBase} d={props} />
+      <DashboardMaintenancePanel d={props} />
+      <DashboardSidecarPanels d={props} />
+      <MemoryObservabilityCard apiBase={props.apiBase} />
+    </>
+  );
+}
