+++ gui/tests/dashboard-contracts.test.ts
@@ -6,14 +6,17 @@ test("project-config diagnostics poll cadence is owned by the shared constant", 
 });
 
 test("Dashboard wires a single project-config diagnostics owner outside the settings poll", async () => {
-  const src = await Bun.file(new URL("../src/pages/Dashboard.tsx", import.meta.url)).text();
-  expect(src.match(/diagnostics\/project-config/g)?.length ?? 0).toBe(1);
-  expect(src).toContain("PROJECT_CONFIG_DIAGNOSTICS_POLL_MS");
-  const fetchDataStart = src.indexOf("const fetchData = async () => {");
-  const fetchDataEnd = src.indexOf("const interval = setInterval(fetchData, 5000);", fetchDataStart);
-  expect(fetchDataStart).toBeGreaterThan(-1);
-  expect(fetchDataEnd).toBeGreaterThan(fetchDataStart);
-  expect(src.slice(fetchDataStart, fetchDataEnd)).not.toContain("diagnostics/project-config");
+  const core = await Bun.file(new URL("../src/pages/dashboard-core-poll.ts", import.meta.url)).text();
+  const hook = await Bun.file(new URL("../src/pages/use-dashboard-data.ts", import.meta.url)).text();
+  // Diagnostics live in their own fetcher + client-resource poll, not inside core health.
+  expect(core.match(/diagnostics\/project-config/g)?.length ?? 0).toBe(1);
+  expect(hook).toContain("fetchProjectConfigDiagnostics");
+  expect(hook).toContain("PROJECT_CONFIG_DIAGNOSTICS_POLL_MS");
+  // Core poll must not own the diagnostics endpoint.
+  const coreFnStart = core.indexOf("export async function fetchDashboardCore");
+  expect(coreFnStart).toBeGreaterThan(-1);
+  const coreBody = core.slice(coreFnStart);
+  expect(coreBody).not.toContain("diagnostics/project-config");
 });
 
 test("Dashboard workspace pane is a labelled section, not a nested main landmark", async () => {
@@ -25,7 +28,7 @@ test("Dashboard workspace pane is a labelled section, not a nested main landmark
 });
 
 test("multi-agent guidance gates injection controls and Active badge on the enabled flag", async () => {
-  const src = await Bun.file(new URL("../src/pages/Dashboard.tsx", import.meta.url)).text();
+  const src = await Bun.file(new URL("../src/pages/dashboard-overview-sections.tsx", import.meta.url)).text();
   expect(src).toContain("!multiAgentGuidanceEnabled");
   expect(src).toContain("multiAgentGuidanceEnabled &&");
   expect(src).toContain("models.v2Mode_");
